+++ CHANGES.rst
@@ -1,6 +1,9 @@
 0.7 (unreleased)
 ----------------
 
+- ``plot_fit`` and ``plot_data`` have new options (``errorbar_opts`` and
+  ``ulim_opts``) to control the properties of spectral flux points and
+  upper-limits.
 
 0.6.1 (2015-10-29)
 ------------------